fix: Properly exit SimpleExecuteQueryStream on stream end
Closes https://github.com/GlareDB/glaredb/issues/2242

Also moves the simple stuff to its own module to split up the code a little bit.
