+++ crates/rpcsrv/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod errors;
 pub mod handler;
 pub mod proxy;
+pub mod simple;
 
 mod session;
